fix URI and literal values not being saved to the database
Because parameter evaluation order is not defined in C, it
was possible (depending on the compiler) that the length
of strings returned by the librdf_*_as_counted_string was
computed too late. As a result 0 was passed as URI or literal
string length to bind_text or to librdf_digest_update.